+++ .gitignore
@@ -6,3 +6,4 @@
 /logs
 /src/admin/.htaccess
 /src/admin/.htpasswd
+/csv
