+++ src/admin/csv.php
@@ -8,7 +8,9 @@ use lib\common\Common;
 use lib\common\PDODatabase;
 use lib\common\Session;
 use lib\common\Token;
-use lib\CSV;
+use lib\InputCSV;
+use lib\ManageUser;
+use lib\OutputCSV;
 use lib\User;
 
 
@@ -67,13 +69,16 @@ if ($_SERVER["REQUEST_METHOD"] == 'POST') {
         $mode = 1;
     } elseif ($_POST['send'] === 'csv_event_register') {
         $mode = 2;
+    } elseif ($_POST['send'] === 'csv_user_output') {
+        $mode = 3;
     }
+
 
     switch ($mode){
         case 1:
             // インスタンス作成
             try {
-                $csv = new CSV($db, $_FILES['csv']);
+                $csv = new InputCSV($db, $_FILES['csv']);
             } catch (\Exception $e) {
                 $err_arr['red__csv_read_failed'] = $e->getMessage();
 
@@ -101,7 +106,7 @@ if ($_SERVER["REQUEST_METHOD"] == 'POST') {
             $ids = User::getIds($db);
             try {
                 $db->dbh->beginTransaction();
-                CSV::insertDummyEvents($db, $ids);
+                InputCSV::insertDummyEvents($db, $ids);
                 $db->dbh->commit();
 
                 $msg_arr['green__dummy_event_registered'] = '入出金データの登録が完了しました。';
@@ -111,6 +116,12 @@ if ($_SERVER["REQUEST_METHOD"] == 'POST') {
                 $msg_arr['red__event_register_failed'] = '入出金データの登録に失敗しました。';
             }
             break;
+
+        case 3:
+            $users = ManageUser::getAllUsersForCSV($db);
+            $csv = new OutputCSV($db);
+            $filepath = $csv->createCSV($users, OutputCSV::OUTPUT_USERS);
+            $csv->downloadCSV($filepath);
     }
 }
