+++ src/lib/InputCSV.class.php
@@ -5,21 +5,20 @@ namespace lib;
 use Google\Service\FirebaseRules\FunctionCall;
 use lib\common\PDODatabase;
 use lib\Category;
+use lib\common\Bootstrap;
 use lib\User;
 use lib\Wallet;
 use Twig\Node\Expression\FunctionExpression;
 
-class CSV {
+class InputCSV
+{
 
-    const OUTPUT_USERS = 0;
-
-    const OUTPUT_EVENTS = 1;
 
     private $db = null;
 
     private $spl = null;
 
-    private static array $err_arr = [];
+    private array $err_arr = [];
 
     /**
      * コンストラクタ
@@ -139,104 +138,6 @@ class CSV {
             $db->insert($table, $insertData);// 2レコード分追加
             $db->insert($table, $insertData);
         }
-    }
-
-    private static function getAllUsersForOutput(PDODatabase $db) : array
-    {
-        $table = ' users ';
-        $column = ' id, user_name, role, email, delete_flg, created_at, updated_at ';
-        $users = $db->select($table, $column);
-
-        return $users;
-    }
-
-    public function downloadCSV(string $filepath) : bool
-    {
-        //【PHP】正しいダウンロード処理の書き方(Qiita)
-        //https://qiita.com/fallout/items/3682e529d189693109eb
-
-        //-- ファイルが読めない時はエラー(もっときちんと書いた方が良いが今回は割愛)
-        if (!is_readable($pPath)) {
-
-        }
-
-        $mimeType = 'text/csv';
-
-        //-- 適切なMIMEタイプが得られない時は、未知のファイルを示すapplication/octet-streamとする
-        if (!preg_match('/\A\S+?\/\S+/', $mimeType)) {
-            $mimeType = 'application/octet-stream';
-        }
-
-        //-- Content-Type
-        header('Content-Type: ' . $mimeType);
-
-        //-- ウェブブラウザが独自にMIMEタイプを判断する処理を抑止する
-        header('X-Content-Type-Options: nosniff');
-
-        //-- ダウンロードファイルのサイズ
-        header('Content-Length: ' . filesize($pPath));
-
-        //-- ダウンロード時のファイル名
-        header('Content-Disposition: attachment; filename="' . basename($pPath) . '"');
-
-        //-- keep-aliveを無効にする
-        header('Connection: close');
-
-        //-- readfile()の前に出力バッファリングを無効化する ※詳細は後述
-        while (ob_get_level()) { ob_end_clean(); }
-
-        //-- 出力
-        readfile($pPath);
-
-        //-- 最後に終了させるのを忘れない
-        exit;
-    }
-
-    /**
-     *
-     */
-    public static function createCSV(array $arr, int $mode, int $user_id = null) : string | false
-    {
-        //ファイル名生成
-        $file_prefix = '';
-        $top_arr = [];
-
-        switch ($mode) {
-            case self::OUTPUT_USERS:
-                $file_prefix = 'users';
-                $top_arr = ['"ID"', '"ユーザー名"', '"管理者フラグ"', '"メールアドレス"', '"削除フラグ"', '"登録日"', '"更新日"'];
-            case self::OUTPUT_EVENTS:
-                $file_prefix = 'events_user-id' . $user_id;
-                $top_arr = [];
-            default:
-                self::$err_arr['mode_invalid'] = '不正なパラメータです。';
-                return false;
-        }
-        array_unshift($arr, $top_arr);
-
-        $filename = date('Y-m-d-H-i-s') . '_' . $file_prefix;
-        $file_dir = '../../csv/';
-        $filepath = self::nameFileWithoutCollision($file_dir, $filename);
-
-        //ファイル作成
-        $fp = fopen($filepath, 'w');
-        foreach ($arr as $line) {
-            fputcsv($fp, $line);
-        }
-        fclose($fp);
-
-    }
-
-    private static function nameFileWithoutCollision(string $file_dir, string $filename) : string
-    {
-        $i = 2;
-        if (file_exists($file_dir . $filename . 'csv')) {
-            while (file_exists($file_dir . $filename . '_' . $i . 'csv')) {
-                $i++;
-            }
-            return $file_dir . $filename . '_' . $i . '.csv';
-        }
-        return $file_dir . $filename . '.csv';
     }
 
     public function getErrArr() : array
