+++ templates/admin/csv.html.twig
@@ -10,31 +10,38 @@
 
 	{% include 'component/error.html.twig' %}
 
-	<h3>CSVユーザー登録</h3>
-	<form action="" method="post" enctype="multipart/form-data">
-		<input type="hidden" name="token" value="{{token}}">
-		<input type="hidden" name="send" value="csv_user_register">
+	<div>
+		<h3>CSVユーザー登録</h3>
+		<form action="" method="post" enctype="multipart/form-data">
+			<input type="hidden" name="token" value="{{token}}">
+			<input type="hidden" name="send" value="csv_user_register">
 
-		<input type="file" name="csv">
-		<button id="user-csv-register-btn" class="btn send_btn" name="send" value="csv_user_register">登録</button>
-	</form>
+			<input type="file" name="csv">
+			<button id="user-csv-register-btn" class="btn send_btn" name="send" value="csv_user_register">登録</button>
+		</form>
+	</div>
+
+	<br>
+	<br>
+
+	<div>
+		<h3>全ユーザーへダミーデータ挿入</h3>
+		<form name="form" action="" method="post">
+			<input type="hidden" name="token" value="{{token}}">
+			<input type="hidden" name="send" value="csv_event_register">
+			<button id="dummy_data_btn" class="btn send_btn" name="send" value="csv_event_register">登録</button>
+		</form>
+	</div>
 
 	<br>
 	<br>
 
-	<h3>全ユーザーへダミーデータ挿入</h3>
-	<form name="form" action="" method="post">
-		<input type="hidden" name="token" value="{{token}}">
-		<input type="hidden" name="send" value="csv_event_register">
-		<button id="dummy_data_btn" class="btn send_btn" name="send" value="csv_event_register">登録</button>
-	</form>
-
-	<br>
-	<br>
-	<h3>CSV出力（ユーザーデータ一覧）</h3>
-	<form action="" method="post">
-		<button class="btn send_btn" name="send" value="csv_user_output">ユーザーデータ出力</button>
-	</form>
+	<div>
+		<h3>CSV出力（ユーザーデータ一覧）</h3>
+		<form action="" method="post">
+			<button class="btn send_btn" name="send" value="csv_user_output">ユーザーデータ出力</button>
+		</form>
+	</div>
 </div>
